refactor: getSuggest tabId if문 -> switch문 변경

+++ server/src/main/java/com/codueon/boostUp/domain/suggest/repository/CustomSuggestRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Pageable;
 
 public interface CustomSuggestRepository {
 
-    Page<GetTeacherSuggest> getTeacherSuggestsOnMyPage(Long lessonId, Long memberId, Long tabId, Pageable pageable);
+    Page<GetTeacherSuggest> getTeacherSuggestsOnMyPage(Long lessonId, Long memberId, int tabId, Pageable pageable);
 
     Page<GetStudentSuggest> getStudentSuggestsOnMyPage(Long memberId, Pageable pageable);
 }
